backend/requirements.txt
alpaca-py==0.13.3
Flask-Migrate==4.0.5
Flask
Flask-SQLAlchemy==3.1.1
requests
openai==0.28
flask-cors
